Basic rendering works

+++ src/app.zig
@@ -1,3 +1,3 @@
 pub fn App(comptime state: type) type {
-    return struct { state: state = null, width: f32 = 100.0, height: f32 = 100.0, title: []const u8 = "App" };
+    return struct { state: state = null, width: u32 = 100.0, height: u32 = 100.0, title: []const u8 = "App" };
 }
